Apps moves to weather_clients
